requirements-common.txt
google-genai==1.52.0